extract campaign readiness routes
