+++ src/api/routers/waves.py
@@ -42,19 +42,12 @@ from src.api.routers.wave_campaign_evidence_routes import (
 from src.api.routers.wave_campaign_read_model_routes import (
     router as campaign_read_model_router,
 )
+from src.api.routers.wave_campaign_readiness_routes import (
+    router as campaign_readiness_router,
+)
 from src.api.routers.wave_route_parameters import (
-    CampaignActiveOnQuery,
-    CampaignEvidenceLimitQuery,
-    CampaignEvidenceOffsetQuery,
     CampaignDefinitionIdPath,
     CampaignDefinitionVersionPath,
-    CampaignIncludeLaunchPackageQuery,
-    CampaignLaunchActorIdOptionalQuery,
-    CampaignLaunchActorIdRequiredQuery,
-    CampaignLaunchCorrelationIdQuery,
-    CampaignLaunchHistoryLimitQuery,
-    CampaignLaunchHistoryOffsetQuery,
-    CampaignLaunchRequestedAsOfDateQuery,
     WaveCorrelationIdHeader,
     WaveCreateIdempotencyKeyHeader,
     WaveIdPath,
@@ -65,13 +58,6 @@ from src.api.routers.wave_campaign_models import (
 )
 from src.api.routers.wave_campaign_launch_http import (
     launch_bulk_review_campaign_definition_response,
-)
-from src.api.routers.wave_campaign_read_http import (
-    get_campaign_definition_launch_package_response,
-    get_campaign_definition_preview_readiness_response,
-    get_campaign_definition_workflow_overview_response,
-    list_campaign_definition_launch_history_response,
-    list_campaign_definition_lifecycle_events_response,
 )
 from src.api.routers.wave_create_preview_http import create_wave_response, preview_wave_response
 from src.api.routers.wave_openapi_examples import (
@@ -100,16 +86,9 @@ from src.core.outcomes.repository import DpmOutcomeReviewRepository
 from src.core.rebalance_runs.service import DpmRunSupportService
 from src.core.waves import (
     DpmBulkReviewCampaignDefinition,
-    DpmBulkReviewCampaignDefinitionLaunchHistoryPage,
-    DpmBulkReviewCampaignDefinitionLaunchPackage,
-    DpmBulkReviewCampaignDefinitionPreviewReadiness,
-    DpmBulkReviewCampaignDefinitionWorkflowOverview,
     DpmBulkReviewCampaignDefinitionRepository,
     DpmWaveReportInput,
     DpmWaveRepository,
-)
-from src.core.waves.campaign_definition_events import (
-    DpmBulkReviewCampaignDefinitionLifecycleEventPage,
 )
 from src.infrastructure.risk_authority import (
     LotusRiskAuthorityClient,
@@ -148,172 +127,7 @@ def get_bulk_review_campaign_definition(
 
 
 router.include_router(campaign_evidence_router)
-
-
-@router.get(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}/lifecycle-events",
-    response_model=DpmBulkReviewCampaignDefinitionLifecycleEventPage,
-    status_code=status.HTTP_200_OK,
-    summary="List bulk-review campaign definition lifecycle events",
-    description=(
-        "Projects bounded lifecycle events for one persisted Manage-owned "
-        "`BulkReviewCampaignDefinition:v1`. Events are derived from the immutable definition "
-        "record and show create, retire, and supersede posture without discovering the global "
-        "portfolio universe, recalculating campaign membership, running maker-checker workflow, "
-        "or claiming OMS execution."
-    ),
-)
-def list_bulk_review_campaign_definition_lifecycle_events(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinitionLifecycleEventPage:
-    return list_campaign_definition_lifecycle_events_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        repository=repository,
-    )
-
-
-@router.get(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}/launch-history",
-    response_model=DpmBulkReviewCampaignDefinitionLaunchHistoryPage,
-    status_code=status.HTTP_200_OK,
-    summary="List bulk-review campaign definition launch history",
-    description=(
-        "Returns a bounded append-only launch audit page for one persisted Manage-owned "
-        "`BulkReviewCampaignDefinition:v1`. The records identify durable waves launched from "
-        "the definition and preserve actor, requested as-of date, correlation, and idempotency "
-        "evidence. They do not imply maker-checker workflow, trade approval, order generation, "
-        "routing, fills, settlement, or OMS execution."
-    ),
-)
-def list_bulk_review_campaign_definition_launch_history(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    limit: CampaignEvidenceLimitQuery = 50,
-    offset: CampaignEvidenceOffsetQuery = 0,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinitionLaunchHistoryPage:
-    return list_campaign_definition_launch_history_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        limit=limit,
-        offset=offset,
-        repository=repository,
-    )
-
-
-@router.get(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}/workflow-overview",
-    response_model=DpmBulkReviewCampaignDefinitionWorkflowOverview,
-    status_code=status.HTTP_200_OK,
-    summary="Get bulk-review campaign workflow overview",
-    description=(
-        "Returns an operator-safe workflow overview for one persisted Manage-owned "
-        "`BulkReviewCampaignDefinition:v1`. The overview composes discovery posture, "
-        "fail-closed preview readiness, lifecycle events, launch history, and optional launch "
-        "package guidance. It does not discover the global portfolio universe, recalculate "
-        "source facts, run maker-checker workflow, approve trades, route orders, or claim OMS "
-        "execution."
-    ),
-)
-def get_bulk_review_campaign_definition_workflow_overview(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    requested_as_of_date: CampaignLaunchRequestedAsOfDateQuery,
-    actor_id: CampaignLaunchActorIdOptionalQuery = None,
-    active_on: CampaignActiveOnQuery = None,
-    include_launch_package: CampaignIncludeLaunchPackageQuery = True,
-    correlation_id: CampaignLaunchCorrelationIdQuery = None,
-    launch_history_limit: CampaignLaunchHistoryLimitQuery = 20,
-    launch_history_offset: CampaignLaunchHistoryOffsetQuery = 0,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinitionWorkflowOverview:
-    return get_campaign_definition_workflow_overview_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        requested_as_of_date=requested_as_of_date,
-        actor_id=actor_id,
-        active_on=active_on,
-        launch_history_limit=launch_history_limit,
-        launch_history_offset=launch_history_offset,
-        include_launch_package=include_launch_package,
-        correlation_id=correlation_id,
-        repository=repository,
-    )
-
-
-@router.get(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}/preview-readiness",
-    response_model=DpmBulkReviewCampaignDefinitionPreviewReadiness,
-    status_code=status.HTTP_200_OK,
-    summary="Check bulk-review campaign definition preview readiness",
-    description=(
-        "Evaluates whether one persisted Manage-owned `BulkReviewCampaignDefinition:v1` can be "
-        "used for new `BULK_REVIEW_CAMPAIGN` preview/create. The response is a bounded "
-        "fail-closed supportability check over lifecycle status, as-of date, source-backed "
-        "candidate eligibility, approval evidence, expiry, and optional actor entitlement. It "
-        "does not create a wave, discover the global portfolio universe, recalculate membership, "
-        "run maker-checker workflow, approve trades, or claim OMS execution."
-    ),
-)
-def get_bulk_review_campaign_definition_preview_readiness(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    requested_as_of_date: CampaignLaunchRequestedAsOfDateQuery,
-    actor_id: CampaignLaunchActorIdOptionalQuery = None,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinitionPreviewReadiness:
-    return get_campaign_definition_preview_readiness_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        requested_as_of_date=requested_as_of_date,
-        actor_id=actor_id,
-        repository=repository,
-    )
-
-
-@router.get(
-    "/campaign-definitions/{campaign_id}/versions/{campaign_version}/launch-package",
-    response_model=DpmBulkReviewCampaignDefinitionLaunchPackage,
-    status_code=status.HTTP_200_OK,
-    summary="Build bulk-review campaign definition launch package",
-    description=(
-        "Builds an operator launch package for one persisted Manage-owned "
-        "`BulkReviewCampaignDefinition:v1`. The package contains fail-closed preview readiness, "
-        "a bounded preview/create request draft, idempotency and correlation headers, and explicit "
-        "operating boundaries for downstream consumers. It does not create a wave, discover the "
-        "global portfolio universe, recalculate membership, run maker-checker workflow, approve "
-        "trades, or claim OMS execution."
-    ),
-)
-def get_bulk_review_campaign_definition_launch_package(
-    campaign_id: CampaignDefinitionIdPath,
-    campaign_version: CampaignDefinitionVersionPath,
-    requested_as_of_date: CampaignLaunchRequestedAsOfDateQuery,
-    actor_id: CampaignLaunchActorIdRequiredQuery,
-    correlation_id: CampaignLaunchCorrelationIdQuery = None,
-    repository: DpmBulkReviewCampaignDefinitionRepository = Depends(
-        get_campaign_definition_repository
-    ),
-) -> DpmBulkReviewCampaignDefinitionLaunchPackage:
-    return get_campaign_definition_launch_package_response(
-        campaign_id=campaign_id,
-        campaign_version=campaign_version,
-        requested_as_of_date=requested_as_of_date,
-        actor_id=actor_id,
-        correlation_id=correlation_id,
-        repository=repository,
-    )
+router.include_router(campaign_readiness_router)
 
 
 @router.post(
